[REFACTOR] Statement.StatementPerformance 클래스를 top level 클래스(StatementPerformance)로 이동
